+++ Nickel/Framework/Patches/WizardPatches.cs
@@ -9,7 +9,7 @@ internal static class WizardPatches
 {
 	internal static EventHandler<GetAssignableStatusesEventArgs>? OnGetAssignableStatuses;
 		
-	private static readonly GetAssignableStatusesEventArgs GetAssignableStatusesEventArgsInstance = new();
+	private static readonly Pool<GetAssignableStatusesEventArgs> GetAssignableStatusesEventArgsPool = new(() => new());
 
 	internal static void Apply(Harmony harmony)
 		=> harmony.Patch(
@@ -20,11 +20,15 @@ internal static class WizardPatches
 
 	private static void GetAssignableStatuses_Postfix(State s, ref List<Status> __result)
 	{
-		var args = GetAssignableStatusesEventArgsInstance;
-		args.State = s;
-		args.Statuses = __result;
-		OnGetAssignableStatuses?.Invoke(null, args);
-		__result = args.Statuses;
+		var result = __result;
+		GetAssignableStatusesEventArgsPool.Do(args =>
+		{
+			args.State = s;
+			args.Statuses = result;
+			OnGetAssignableStatuses?.Invoke(null, args);
+			result = args.Statuses;
+		});
+		__result = result;
 	}
 
 	internal sealed class GetAssignableStatusesEventArgs